+++ src/main/java/com/atwoz/survey/domain/membersurvey/dto/SurveyQuestionSubmitCreateDto.java
@@ -1,0 +1,7 @@
+package com.atwoz.survey.domain.membersurvey.dto;
+
+public record SurveyQuestionSubmitCreateDto(
+        Long questionId,
+        Integer answerNumber
+) {
+}
